+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.0
+**Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
+**Date**: Fri Mar 09 10:39:05 CET 2018
+
+* [FEATURE] Configure services of the same type via parent classes or interfaces
+* [CHANGE] Exracted configuration into Micro\Container\Config
+
+
 ## 0.1.4
 **Maintainer**: Raffael Sahli <sahli@gyselroth.com>\
 **Date**: Fri Feb 23 13:34:01 CET 2018